remove version from vcap_common in Gemfile
This was removed upstream, but our merge must have conflicted
and this was not removed.

This meant we were stuck on an older incompatiable
vcap_common and many tests were failing due to this.

Need to set some version of commit id in ccng Gemfile upstream
and submit pull request. Otherwise it's impossible to know
when a bundle update is required.