Pushes: update the dea ads synchronously to avoid race conditions with dea_ng
JIRA: https://activestate.atlassian.net/browse/STO-856
Bugzilla: https://bugs.activestate.com/show_bug.cgi?id=104474

The problem is that on code-updates, the client uploads the new bits,
then tells the CC to stop the targeted app, and then tells it to
start (a restart). The CC uses the current advertisements from the
DEA to determine if there's enough space.  The DEA sends this info
to the CC every config.dea_ng.intervals.advertise seconds.  This
change guarantees that the CC will receive at least one notification
between the times the app is stopped and the time it attempts to
restart it.