[105333][STO-1158] Merge branch '105333-cc-monitoring'
Add new monitoring job to watch cloud controller's memory usage and
restart if it it's using too much (exactly how much is configurable).
This is a wallpaper over the fact that the cloud controller seems to be
leaking memory at some point past twenty-four hours of uptime.  The
hope is that upstream will fix the underlying leak in the future and we
will pick up the change.