[105668] Don't change app version upon initial droplet upload
The staging request includes a start request for the first instance of the app.  If the app
version changes once the droplet has been uploaded to the blob store, then the HM will send
a stop request to this initial instance and start a new instance at the changed version,
which would actually be identical to the initial version.