[105573] Make sure that service credentials are sent over mbus as a hash.
From what I can see upstream has changed the storage of credentials dramatically against
what we have currently so this is a workaround to make sure that credentials are sent
over the mbus as a hash now that they are stored as a string.

See commit dc6df0a and 9ec7ccd for examples of upstream changing on service binding.