300170 - Make test suite work on any linux
The spec/ suite needed to be running on a stackato instance. Now it can run
anywhere by anyone using a setup script and a runner script.

The setup script (tested on a plain Ubuntu 14.04 vm) is here:

    spec/stackato/setup-test-env.sh

And the test runner script is here:

    spec/stackato/run-spec-tests.sh

The runner sets up some env vars and runs:

    bundle exec rspec

The set up instructions are here:

    spec/README.md

The spec needs a lot of things to be setup, but the script will (hopefully)
catch them all and tell you what to do next. Just keep running over and over
until it works.

The tests are currently failing for me on this error:

    https://gist.github.com/d6dbaac92558c4683cc8

which needs to be looked into. This is a starting point.

This can be used to get the tests into CI.