300037 Support staging and starting docker apps
Docker apps don't include any uploaded application bits; they are
specified by a docker image name:

$ stackato push -n --as myapp --docker-image mycompany/mayapp:tag

Each CC will host a docker registry (with shared image storage). The
"staging" of a docker app consists of pulling the image from an
external registry and pushing it to an internal one.  The stager will
send back the docker image id to the CC, which will turn it in the
droplet_hash field in the database.

The application start message includes both the docker image name, and
the image id.  The name itself is not used once the image has been
copied to the internal registry, but the fact that the name isn't nil
indicates that droplet_hash must be treated as a docker id and not as
a droplet hash when starting a new app instance.