301378 Unbreak credentials of v1 services when updated without instances
When we see an update on /services/v1/offerings/*/handles/:id, and there is
no active ManagedServiceInstance with that id, we directly update the
service binding.  When doing so, we need to pass in the credentials as a
hash to be serialized by ServiceBinding#credentials_with_serialization=
so that it can be deserialized correctly.